refactor: update types names/definitions for easier dx
