+++ src/comment/models/get-comments-response.model.ts
@@ -1,0 +1,13 @@
+interface RawCommentsResponseEntity {
+  Count: number;
+  Total: number;
+  Results: RawComment[];
+}
+
+interface RawComment {
+  User: string;
+  Submitted: string;
+  CommentText: string;
+}
+
+export type GetCommentsResponse = RawCommentsResponseEntity;
