+++ src/comment/models/index.ts
@@ -1,2 +1,3 @@
-export * from "./comments.model";
+export * from "./comment-entity.model";
+export * from "./comment-response.model";
 export * from "./get-comments-response.model";
